decomposed script module;
