+++ src/anim_state_control.cpp
@@ -10,6 +10,7 @@
 #include "render/camera.h"
 #include "render/frustum.h"
 #include "render/render.h"
+#include "script/script.h"
 #include "engine.h"
 #include "audio.h"
 #include "physics.h"
@@ -18,7 +19,6 @@
 #include "game.h"
 #include "skeletal_model.h"
 #include "entity.h"
-#include "script.h"
 #include "anim_state_control.h"
 #include "character_controller.h"
 #include "vt/tr_versions.h"
