+++ src/audio.cpp
@@ -31,10 +31,10 @@ extern "C" {
 #include "core/vmath.h"
 #include "render/camera.h"
 #include "render/render.h"
+#include "script/script.h"
 #include "vt/vt_level.h"
 #include "game.h"
 #include "audio.h"
-#include "script.h"
 #include "engine.h"
 #include "entity.h"
 #include "character_controller.h"
