+++ src/engine.cpp
@@ -25,6 +25,7 @@ extern "C" {
 #include "core/gl_text.h"
 #include "render/camera.h"
 #include "render/render.h"
+#include "script/script.h"
 #include "vt/vt_level.h"
 #include "game.h"
 #include "audio.h"
@@ -36,7 +37,6 @@ extern "C" {
 #include "room.h"
 #include "world.h"
 #include "resource.h"
-#include "script.h"
 #include "engine.h"
 #include "physics.h"
 #include "controls.h"
