+++ src/entity.cpp
@@ -12,6 +12,7 @@ extern "C" {
 #include "core/obb.h"
 #include "render/camera.h"
 #include "render/render.h"
+#include "script/script.h"
 #include "vt/tr_versions.h"
 #include "audio.h"
 #include "mesh.h"
@@ -21,7 +22,6 @@ extern "C" {
 #include "world.h"
 #include "engine.h"
 #include "physics.h"
-#include "script.h"
 #include "trigger.h"
 #include "anim_state_control.h"
 #include "character_controller.h"
