+++ src/gui.cpp
@@ -14,6 +14,7 @@
 #include "render/render.h"
 #include "render/shader_description.h"
 #include "render/shader_manager.h"
+#include "script/script.h"
 #include "image.h"
 #include "gui.h"
 #include "mesh.h"
@@ -22,7 +23,6 @@
 #include "character_controller.h"
 #include "engine.h"
 #include "audio.h"
-#include "script.h"
 #include "engine_string.h"
 #include "world.h"
 #include "inventory.h"
