+++ src/physics_bullet.cpp
@@ -24,11 +24,11 @@ extern "C" {
 #include "core/console.h"
 #include "core/obb.h"
 #include "render/render.h"
+#include "script/script.h"
 #include "engine.h"
 #include "mesh.h"
 #include "skeletal_model.h"
 #include "character_controller.h"
-#include "script.h"
 #include "entity.h"
 #include "resource.h"
 #include "room.h"
