+++ src/render/render.cpp
@@ -11,6 +11,7 @@
 #include "../core/vmath.h"
 #include "../core/polygon.h"
 #include "../core/obb.h"
+#include "../script/script.h"
 #include "../vt/tr_versions.h"
 #include "camera.h"
 #include "render.h"
@@ -20,7 +21,6 @@
 #include "shader_manager.h"
 #include "../room.h"
 #include "../world.h"
-#include "../script.h"
 #include "../mesh.h"
 #include "../skeletal_model.h"
 #include "../entity.h"
