+++ src/script/script.h
@@ -8,6 +8,11 @@ struct lua_State;
 
 #define CVAR_LUA_TABLE_NAME "cvars"
 
+// Response constants
+#define RESP_KILL           (0)
+#define RESP_VERT_COLLIDE   (1)
+#define RESP_HOR_COLLIDE    (2)
+#define RESP_SLIDE          (3)
 
 extern lua_State *engine_lua;
