+++ src/trigger.cpp
@@ -13,9 +13,9 @@ extern "C" {
 #include "core/system.h"
 #include "core/console.h"
 #include "core/vmath.h"
+#include "script/script.h"
 #include "room.h"
 #include "trigger.h"
-#include "script.h"
 #include "gameflow.h"
 #include "game.h"
 #include "audio.h"
