+++ src/world.cpp
@@ -21,6 +21,7 @@ extern "C" {
 #include "render/frustum.h"
 #include "render/render.h"
 #include "render/bordered_texture_atlas.h"
+#include "script/script.h"
 #include "vt/vt_level.h"
 #include "audio.h"
 #include "room.h"
@@ -32,7 +33,6 @@ extern "C" {
 #include "anim_state_control.h"
 #include "engine.h"
 #include "physics.h"
-#include "script.h"
 #include "gui.h"
 #include "gameflow.h"
 #include "resource.h"
